# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/client/{clientId}/address": {
      "put": {
        "tags": [
          "client-controller"
        ],
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/client": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/locale/states": {
      "get": {
        "tags": [
          "locales-controller"
        ],
        "operationId": "getStates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/State"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/locale/municipalities/{stateId}": {
      "get": {
        "tags": [
          "locales-controller"
        ],
        "operationId": "getMunicipalitiesByStateID",
        "parameters": [
          {
            "name": "stateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Municipality"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/locale/address/{cep}": {
      "get": {
        "tags": [
          "locales-controller"
        ],
        "operationId": "getClientByCPF",
        "parameters": [
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cep"
                }
              }
            }
          }
        }
      }
    },
    "/client/{cpf}": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClientByCPF_1",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/client/id/{id}": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClientByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddressRequestDTO": {
        "required": [
          "cep",
          "city",
          "number",
          "state",
          "street"
        ],
        "type": "object",
        "properties": {
          "cep": {
            "maxLength": 8,
            "minLength": 8,
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "maxLength": 2,
            "minLength": 2,
            "type": "string"
          }
        }
      },
      "ClientRequestDTO": {
        "required": [
          "address",
          "cpf",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cpf": {
            "maxLength": 11,
            "minLength": 11,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "address": {
            "$ref": "#/components/schemas/AddressRequestDTO"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "cep": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "ClientResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cpf": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "dateOfBirth": {
            "type": "string"
          }
        }
      },
      "State": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "acronym": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Municipality": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
      "Cep": {
        "type": "object",
        "properties": {
          "cep": {
            "type": "string"
          },
          "patio": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      }
    }
  }
}